+++ src/ecs/spawn.h
@@ -5,3 +5,4 @@
 void spawn_normal_boids(flecs::world& world, int count);
 void spawn_doctor_boids(flecs::world& world, int count);
 void spawn_initial_population(flecs::world& world);
+void reset_simulation(flecs::world& world);